COM-1404-final add tests
