+++ tests/integration/programs.test.js
@@ -330,89 +330,6 @@ describe('PROGRAMS ROUTES - POST /programs/{_id}/subprogram', () => {
   });
 });
 
-describe('PROGRAMS ROUTES - POST /programs/{_id}/step', () => {
-  let authToken = null;
-  beforeEach(populateDB);
-  const payload = { name: 'new step', type: 'e_learning' };
-
-  describe('VENDOR_ADMIN', () => {
-    beforeEach(async () => {
-      authToken = await getToken('vendor_admin');
-    });
-
-    it('should create step', async () => {
-      const programId = programsList[0]._id;
-      const stepsLengthBefore = programsList[0].steps.length;
-      const response = await app.inject({
-        method: 'POST',
-        url: `/programs/${programId.toHexString()}/steps`,
-        payload,
-        headers: { 'x-access-token': authToken },
-      });
-
-      const programUpdated = await Program.findById(programId);
-
-      expect(response.statusCode).toBe(200);
-      expect(programUpdated._id).toEqual(programId);
-      expect(programUpdated.steps.length).toEqual(stepsLengthBefore + 1);
-    });
-
-    const missingParams = ['name', 'type'];
-    missingParams.forEach((param) => {
-      it(`should return a 400 if missing ${param}`, async () => {
-        const programId = programsList[0]._id;
-        const response = await app.inject({
-          method: 'POST',
-          url: `/programs/${programId.toHexString()}/steps`,
-          payload: omit(payload, param),
-          headers: { 'x-access-token': authToken },
-        });
-
-        expect(response.statusCode).toBe(400);
-      });
-    });
-
-    it('should return a 400 if program does not exist', async () => {
-      const wrongId = new ObjectID();
-      const response = await app.inject({
-        method: 'POST',
-        url: `/programs/${wrongId}/steps`,
-        payload,
-        headers: { 'x-access-token': authToken },
-      });
-
-      expect(response.statusCode).toBe(400);
-    });
-  });
-
-  describe('Other roles', () => {
-    const roles = [
-      { name: 'helper', expectedCode: 403 },
-      { name: 'auxiliary', expectedCode: 403 },
-      { name: 'auxiliary_without_company', expectedCode: 403 },
-      { name: 'coach', expectedCode: 403 },
-      { name: 'client_admin', expectedCode: 403 },
-      { name: 'training_organisation_manager', expectedCode: 200 },
-      { name: 'trainer', expectedCode: 403 },
-    ];
-
-    roles.forEach((role) => {
-      it(`should return ${role.expectedCode} as user is ${role.name}`, async () => {
-        authToken = await getToken(role.name);
-        const programId = programsList[0]._id;
-        const response = await app.inject({
-          method: 'POST',
-          payload,
-          url: `/programs/${programId.toHexString()}/steps`,
-          headers: { 'x-access-token': authToken },
-        });
-
-        expect(response.statusCode).toBe(role.expectedCode);
-      });
-    });
-  });
-});
-
 describe('POST /programs/:id/cloudinary/upload', () => {
   let authToken;
   let form;
