+++ tests/integration/subProgram.test.js
@@ -1,5 +1,7 @@
 const expect = require('expect');
 const app = require('../../server');
+const { ObjectID } = require('mongodb');
+const omit = require('lodash/omit');
 const SubProgram = require('../../src/models/SubProgram');
 const { populateDB, subProgramsList } = require('./seed/subProgramsSeed');
 const { getToken } = require('./seed/authenticationSeed');
@@ -66,6 +68,89 @@ describe('SUBPROGRAMS ROUTES - PUT /subprograms/{_id}', () => {
           method: 'PUT',
           payload,
           url: `/subprograms/${subProgramId.toHexString()}`,
+          headers: { 'x-access-token': authToken },
+        });
+
+        expect(response.statusCode).toBe(role.expectedCode);
+      });
+    });
+  });
+});
+
+describe('SUBPROGRAMS ROUTES - POST /subprograms/{_id}/step', () => {
+  let authToken = null;
+  beforeEach(populateDB);
+  const payload = { name: 'new step', type: 'e_learning' };
+
+  describe('VENDOR_ADMIN', () => {
+    beforeEach(async () => {
+      authToken = await getToken('vendor_admin');
+    });
+
+    it('should create step', async () => {
+      const subProgramId = subProgramsList[0]._id;
+      const stepsLengthBefore = subProgramsList[0].steps.length;
+      const response = await app.inject({
+        method: 'POST',
+        url: `/subprograms/${subProgramId.toHexString()}/steps`,
+        payload,
+        headers: { 'x-access-token': authToken },
+      });
+
+      const programUpdated = await SubProgram.findById(subProgramId);
+
+      expect(response.statusCode).toBe(200);
+      expect(programUpdated._id).toEqual(subProgramId);
+      expect(programUpdated.steps.length).toEqual(stepsLengthBefore + 1);
+    });
+
+    const missingParams = ['name', 'type'];
+    missingParams.forEach((param) => {
+      it(`should return a 400 if missing ${param}`, async () => {
+        const programId = subProgramsList[0]._id;
+        const response = await app.inject({
+          method: 'POST',
+          url: `/subprograms/${programId.toHexString()}/steps`,
+          payload: omit(payload, param),
+          headers: { 'x-access-token': authToken },
+        });
+
+        expect(response.statusCode).toBe(400);
+      });
+    });
+
+    it('should return a 400 if program does not exist', async () => {
+      const wrongId = new ObjectID();
+      const response = await app.inject({
+        method: 'POST',
+        url: `/subprograms/${wrongId}/steps`,
+        payload,
+        headers: { 'x-access-token': authToken },
+      });
+
+      expect(response.statusCode).toBe(400);
+    });
+  });
+
+  describe('Other roles', () => {
+    const roles = [
+      { name: 'helper', expectedCode: 403 },
+      { name: 'auxiliary', expectedCode: 403 },
+      { name: 'auxiliary_without_company', expectedCode: 403 },
+      { name: 'coach', expectedCode: 403 },
+      { name: 'client_admin', expectedCode: 403 },
+      { name: 'training_organisation_manager', expectedCode: 200 },
+      { name: 'trainer', expectedCode: 403 },
+    ];
+
+    roles.forEach((role) => {
+      it(`should return ${role.expectedCode} as user is ${role.name}`, async () => {
+        authToken = await getToken(role.name);
+        const programId = subProgramsList[0]._id;
+        const response = await app.inject({
+          method: 'POST',
+          payload,
+          url: `/subprograms/${programId.toHexString()}/steps`,
           headers: { 'x-access-token': authToken },
         });
