+++ tests/unit/helpers/steps.test.js
@@ -1,7 +1,7 @@
 const sinon = require('sinon');
 const expect = require('expect');
 const { ObjectID } = require('mongodb');
-const Program = require('../../../src/models/Program');
+const SubProgram = require('../../../src/models/SubProgram');
 const Step = require('../../../src/models/Step');
 const StepHelper = require('../../../src/helpers/steps');
 require('sinon-mongoose');
@@ -34,45 +34,45 @@ describe('updateStep', () => {
 });
 
 describe('addStep', () => {
-  let ProgramMock;
+  let SubProgramMock;
   let StepMock;
 
   beforeEach(() => {
-    ProgramMock = sinon.mock(Program);
+    SubProgramMock = sinon.mock(SubProgram);
     StepMock = sinon.mock(Step);
   });
 
   afterEach(() => {
-    ProgramMock.restore();
+    SubProgramMock.restore();
     StepMock.restore();
   });
 
-  const program = { _id: new ObjectID() };
+  const subProgram = { _id: new ObjectID() };
   const newStep = { name: 'c\'est une étape !', type: 'lesson' };
   it('should create a step', async () => {
     const stepId = new ObjectID();
-    ProgramMock.expects('countDocuments').withExactArgs({ _id: program._id }).returns(1);
+    SubProgramMock.expects('countDocuments').withExactArgs({ _id: subProgram._id }).returns(1);
 
     StepMock.expects('create').withExactArgs(newStep).returns({ _id: stepId });
 
-    ProgramMock.expects('updateOne').withExactArgs({ _id: program._id }, { $push: { steps: stepId } });
+    SubProgramMock.expects('updateOne').withExactArgs({ _id: subProgram._id }, { $push: { steps: stepId } });
 
-    await StepHelper.addStep(program._id, newStep);
+    await StepHelper.addStep(subProgram._id, newStep);
 
-    ProgramMock.verify();
+    SubProgramMock.verify();
     StepMock.verify();
   });
 
   it('should return an error if program does not exist', async () => {
     try {
-      ProgramMock.expects('countDocuments').withExactArgs({ _id: program._id }).returns(0);
+      SubProgramMock.expects('countDocuments').withExactArgs({ _id: subProgram._id }).returns(0);
 
       StepMock.expects('create').never();
-      ProgramMock.expects('updateOne').never();
+      SubProgramMock.expects('updateOne').never();
 
-      await StepHelper.addStep(program._id, newStep);
+      await StepHelper.addStep(subProgram._id, newStep);
     } catch (e) {
-      ProgramMock.verify();
+      SubProgramMock.verify();
       StepMock.verify();
     }
   });
